tests/wpt/tests/html/canvas/element/path-objects/2d.path.isPointInPath.subpath.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.isPointInPath.subpath</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.isPointInPath.subpath</h1>
<p class="desc">isPointInPath() uses the current path, not just the subpath</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("isPointInPath() uses the current path, not just the subpath");
_addTest(function(canvas, ctx) {

  ctx.rect(0, 0, 20, 20);
  ctx.beginPath();
  ctx.rect(20, 0, 20, 20);
  ctx.closePath();
  ctx.rect(40, 0, 20, 20);
  _assertSame(ctx.isPointInPath(10, 10), false, "ctx.isPointInPath(10, 10)", "false");
  _assertSame(ctx.isPointInPath(30, 10), true, "ctx.isPointInPath(30, 10)", "true");
  _assertSame(ctx.isPointInPath(50, 10), true, "ctx.isPointInPath(50, 10)", "true");

});
</script>
